Create playground component
